+++ src/app/feature/registry/acsx290-list-control/acsx290-list-control.model.ts
@@ -9,7 +9,9 @@ export interface ACSx290ListControlModel {
   baseDbId: string;
   baseDb: string;
   addendum: string;
+  procedureDateTime: string;
   completion: number;
   tags: { tag: string; priority: string }[];
+  createdAt: any;
   modifiedAt: any;
 }