+++ src/app/feature/registry/cath-pci50-list-control/cath-pci50-list-control.model.ts
@@ -8,8 +8,10 @@ export interface CathPci50ListControlModel {
   baseDbId: string;
   baseDb: string;
   addendum: string;
+  procedureDateTime: string;
   completion: number;
   tags: { tag: string; priority: string }[];
   submitted: { submit: string; label: string; endpoint: string; priority: string }[];
+  createdAt: any;
   modifiedAt: any;
 }