+++ src/environments/environment.ts
@@ -14,6 +14,15 @@ export const environment = {
     messagingSenderId: '543502718120',
     appId: '1:543502718120:web:bdc130d783417244'
   }
+  // firebase: {
+  //   apiKey: 'AIzaSyC3kGI6Ht4l5T31WgafOLRtlzM9e7fVdSo',
+  //   authDomain: 'bdms-registry.firebaseapp.com',
+  //   databaseURL: 'https://bdms-registry.firebaseio.com',
+  //   projectId: 'bdms-registry',
+  //   storageBucket: 'bdms-registry.appspot.com',
+  //   messagingSenderId: '383870724152',
+  //   appId: '1:383870724152:web:2ffda41defbb1215'
+  // }
 };
 
 /*